Formatted Update
Martin,
The only thing that I can figure out is that at some point I must have accidentally keyed a "Ctrl-V" instead of a "Ctrl-C".
So I started over completely and copied the latest file from your master and went through it this morning. I have not touched this file except for the formatting and it is all white-space changes.
Thanks for your patience with me.
I loaded this into my AWS and ran teats on it. Everything I tried worked and the logs were clean.
Thanks Again,
Wes